[fix][change] do not make cons an orthogonally specializing type
If a type specifier depends on another type specifier that is
not a tag stored within some object, then there is no unique
way to determine what such a type specifier could be. In particular,
an object that is of type (cons fixnum fixnum) could also be of type
(cons integer integer) (cons bit bit) or (cons fixnum integer) or any
of the other infinite variants. Thus, it does not make sense (however
appealing) to treat cons as an orthogonally specializing type.